Issue 52. Make the output of "muddle depend" more concise.
Each dependency is now output on a single line. Whilst the lines can get
long, this is still easier to read than before. It also allows one to
"grep" the output to select lines of interest (which was not possible
before).

Rules which have no depenencies can now be suppressed by adding "-short"
to the "all", "system" or "user" keyword.

NB: Added __cmp__ and __hash__ to the Rule class, but I believe it still
works.


git-svn-id: https://muddle.googlecode.com/svn/trunk/muddle@191 837abeb0-121e-11de-ab22-55a8a853a75f