Make Bash script work with newer Ubuntu
Current versions of Ubuntu seem to keep git contributed hooks
in a slightly different place. Make the Bash script aware of this.

Also, change the script so it uses the contrib hook via a direct
link, rather than copying it to /usr/local/lib first. And make
it aware of previously possible values (to an extent) so it will
"upgrade" them (with distinct messages for each action it actually
performs).