Rethink handling of _all
The in-code documentation as it stands says _all means what it says
- all of the applicable type of label.

Trying to allow constructs like _all{role} or (domain)_all are
confusing. The latter is especially confusing, because there is
no ready way to say "all the labels in the top level domain",
given that _all means *all* labels, and (domain)_all means
all labels in (domain). Trying to allow ()_all is nasty, and
rapidly gets more confusing. Anyway, if (domain)_all meant all
the labels in domain (domain), why *should* _all mean *all* labels,
when normal Label nomenclature says it should only mean the top
level.

Clearly the answer is to retract complicated use of _all. Keep
it simple. If we need wildcarding at a more complicated level,
support use of '*' to mean such. Although, of course, that may
cause confusion if we've got labels which are *named* with stars
as part of their representation.