Experimenting with NOT always specifying default_domain
Now we're better at passing real labels, this may be sensible.

Also, trying to guess a domain is awkward - conside if there is a
checkout called FRED in both toplevel and subdomains - if we're in
the subdomain, and we are willing to "guess" that as our domain,
how are we meant to specify we actually want the toplevel FRED
(there's no way to do it with a label if the default domain is
going to be inserted for us, as we don't have a denotation for the
toplevel domain, as always writing () would be awful, and *sometimes*
writing it is not much better).