Try to explain when nothing happens
Add code to try to explain why nothing happens when a CPD command
finds nothing to do. This makes the code a bit more complex, but
should hopefully reduce the amount of starting at things wondering
WHY nothing has happend when the user thought they gave a perfectly
good command (especially in the case where the problem is that a
package name is not used because the package exists but is not in
a default role).