Make commands that fail with a problem list GiveUp
That is, if a command such as "muddle push" reports it has failed
to push several checkouts, then make it raise GiveUp, so that the
muddle command itself will exit with status code 1 (to the shell).

This allows test scripts (which run muddle as a command) to stop
when they hit such a problem.

In order to facilitate this, make __main__.py not print out anything
for an empty GiveUp exception.