Stop "muddle reparent" grumbling with git
Specifically, if there wasn't a "remote origin" entry in .git/config,
then muddle reparent would cause worrying sounding messages when it
did "git remote rm origin" (because that would grumble there was
nothing to remove, and it LOOKED like an error message).

So check if there's anything to remove first.