First pass at refactoring of distribution
This splits "binary" package distribution up into "obj" and "install"
distribution, allows passing a package label to distribute_checkout,
allows package labels to distribute_checkout and distribute_package
to be wildcarded (and thus we also remove distribute_role), tries
to find the Makefile.muddle for each package and default to
distributing that as well, and makes _binary_release be an "install"
distribution.

All of which is untested, and moreover the test_distribute code has
not yet been updated, so will fail magnificantly.

Also, the documentaiton is now out-of-date.