Put a blank line before a muddle error
If muddle exists with a GiveUp or MuddleBug, put a blank line
before the error. This may cause an extra unneeded blank line
in some cases, but in many cases it will be the only blank line
being added, and in those cases I think it makes it a lot easier
to see where the error message starts.

(One could argue for something more than a blank line, but that's
a discussion for another time).