Got stderr folding into stdout properly now
Previously, for some commands, stderr and stdout were not folding
together correctly - specifically, the stderr text was coming out
first.

This is presumably because the underlying command (git?) was buffering
differently when output was to a terminal (so the command reported its
output sanely when run at the prompt) versus when outputting to an
internal buffer (or pipe).

The solution is to ask for buffering to be disabled, via an environment
variable, so it propagates downwards.

(setting the bufsize argument to subprocess calls will not have the
same effect)