Handle labels properly inside db.py
1. Add _inner_labels to the db.py Database class, to return all
   the labels it knows about.
2. Make mechanics.py use that to set the domain for labels when
   including a domain
3. And thus the include_domain method in Database can be a lot
   simpler

Because we're now doing labels-in-a-subdomain properly in db.py,
we can now amend normalise_checkout_label so that it does not
make a copy of the label unless it needs to. Which should reduce
the number of labels used in a build tree by some significant
amount.

Also, whilst doing that, make a VCS checkout action have its
own _inner_labels method, which just returns the result of the
inner vcs._inner_labels() method (as it should).