Internally, rename "fetch" to "pull"
This has been bugging me for a while, since "muddle pull" is preferred,
but internally the VCS commands all used "fetch" as the term-of-choice,
and that keeps confusing me.

So I've given in and fixed it.

NB: The Fetched tag (/fetch) has always been transient, so should
never occur (persistently) in a .muddle/tags/checkout/xxx directory,
so it shouldn't matter that this is now a Pushed tag (/pushed).