Finally remove 'builder' from VersionControlHandler
We've known for a long time that the (seemingly sensible) decision
to put a 'builder' value inside a VersionControlHandler, rather than
passing it run time as everything else does, was a mistake. It was a
mistake because if one is in a build that becomes a subdomain, the
"bound in" 'builder' is the one for the subdomain, and by the time we
want to *use* the VersionControlHandler, the builder we're actually
wanting to use (for everything) is the one for the top-level domain.

Getting round this required some *odd* stuff in the include_domain()
code in mechanics.py. The only remaining oddness is needing to remember
to "pass up" the VersionControlHandler()._inner_labels() list of labels
when scanning a (soon to be) subdomain for labels to translate. Which is
still a bit messy, but it's the same *sort* of messiness we might
expect, rather than a different kind of oddness.