Pull some changes from the lifecycle development
In order to allow "muddle pull" to specify a particular revision
(in git, anyway) we need to pull some changes across from the
lifecycle-* branches, which are currently in development.

This isn't as good as waiting for that development to be finished
and using the result, but it should be enough to get us going.