"Deploy" to the package install directory.
If we install to the obj/<name>/<role> directory, then it is not
possible to immediately "grab" the CPIO file with a deployment
(since deployment actions look in the install directoriees).

If we install to the install/<role> directory, then that's solved,
but a *package* wanting to use the CPIO file (for instance, to do
a kernel build) can still access it in the install directory just
as easily.

Also, given that the CPIO filename can include a subdirectory,
it is quite possible to use the same role for all CPIO package builds,
just with different subdirectories.