Rename the db.py PathFile values, add desc branch
Rename the PathFile values inside Database so that it is obvious
what they are, and so the names relate more directly to the
filenames we use in the .muddle/ directory.

Add *optional* DescriptionBranch, which can be used to record
the (top level build description) branch from a "muddle init
-branch" command. Thus also add PathFile.get_if_it_exists().

Amend version stamps to record this value if it exists
(but not if it doesn't). This still needs some extra tests
adding to the VersionStamp tests. Whilst this is a change
to the data in a stamp file, I don't think it's enough to
warrant incrementing the stamp file version number.

Amend the Stamp/Unstamp support to recognise the description
branch if it is given.

Note that we do not, at the moment, provide any way of doing
the same sort of thing for build descriptions in subdomains.