Fixing subdomain support, testing various things
When doing a "muddle init" with subdomains and a toplevel "following"
build description, "sync" the subdomain build descriptions at the
end of the "muddle init" (which is the only feasible place to do this).

Cope with checkouts that are not checked out (well, can't find their
current branch) in "muddle query checkout-branches" - this allows
using it just after "muddle init", for instance.

Add tests for a "following" build description with subdomains,
and some testing for remote/upstream repositories still working
as they should (both in test_lifecycle2.py)