"muddle bootstrap" now does "muddle reparent"
Well, actually it loads the build description (which it didn't used
to do, and is probably a useful check in and of itself), and then
does a reparent. This fixes issue 210.

Various test scripts then:

1. Do not need to do a muddle reparent after doing a bootstrap
2. Do need to delete 01.pyc, where they previously did not, because
   muddle bootstrap itself will have created it (by loading the
   build description)

I've also done some cosmetic changes to the test_checkout_bzr.py
script, to use muddle import rather than muddle assert, and also
to show using muddle push as well as bzr push for new checkouts.