Clarify some comments.
I wanted to be sure why it was necessary for "muddle unstamp -u"
to use "muddle pull -reload" (see around line 4819).

So when we're doing an update from a stamp file, we take our
initial build tree (that of the build tree as it was before
our "muddle unstamp -u" command, and:

1. Add any subdomain directories that are missing
2. Add any checkouts (to the 'builder') that are missing
3. Amend the repository information for any checkouts that
   have changed their repository.
4. Set specific revision ids for any checkouts whose revision
   ids have changed.

We then use "muddle pull" *of that hand assembled build tree"
to make everything become consistent with it (that is, we don't
re-read the build description and pull, we use the build tree
we've amended).

If we used "muddle pull" without the "-noreload" flag, then
"muddle pull" might pull the build description, and if that
changed, then "muddle pull" would *reload" the build description,
and throw away our hand-amended 'builder' build tree.

Since we don't want that, we should use "-noreload".

(It is implicit that if the build description does change,
by being pulled, it is being changed towards the state that
was used to write the stamp file, so we should already have
in hand all the changes needed to match the updated build
description.)