Complain if a spurious .muddle is found
For some reason my src/builds directory contained an empty directory
called .muddle. Of course, I couldn't see this, but if I tryed to
use any "muddle" command (that needed a build tree) in that directory,
it exploded at me in an obscure manner.

Whilst it's quite unlikely that people will encounter this if they're
not doing strange muddle development things (I hope), I've added this
code so that at least some sort of sensible message is output
explaining the problem, and so I won't think it's a bug in muddle
itself.