Investigating output
The output produced now is, I think, better than previously,
as determined by doing:

  <muddle> all_tests.py &> ../../<muddle>.test.txt

for each of <muddle> being muddle and m3. and then editing
m3.test.txt to change "m3" to "muddle" throughout, and then
comparing the two files side-by-side.

There are stil places where m3 (this version) is losing output,
but it's less than with the older (master branch) muddle, and I think
that on the whole output is interleaved in a much better manner.

So some investigation still needed, and we *really* want to be able
to do "muddle -o <output-file>" to have ALL output go to <output-file>
so I don't have to do the "&> <file>" thing.